Provide an assembler for WPG dataset
