+++ jena-arq/src/main/java/org/apache/jena/sparql/core/assembler/WriterPerGraphDatasetAssembler.java
@@ -1,0 +1,41 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.jena.sparql.core.assembler;
+
+import org.apache.jena.assembler.Assembler;
+import org.apache.jena.query.Dataset;
+import org.apache.jena.query.DatasetFactory;
+import org.apache.jena.rdf.model.Resource;
+import org.apache.jena.sparql.core.DatasetGraphPerGraphLocking;
+
+/**
+ * An {@link Assembler} that creates writer-per-graph in-memory {@link Dataset}s.
+ */
+public class WriterPerGraphDatasetAssembler extends TransactionalInMemDatasetAssembler {
+
+    @Override
+    public Resource getType() {
+        return DatasetAssemblerVocab.tDatasetPGraphWriter ;
+    }
+    
+    @Override
+    public Dataset createDataset() {
+        return DatasetFactory.wrap(new DatasetGraphPerGraphLocking()) ;
+    }
+}
