+++ jena-arq/src/test/java/org/apache/jena/sparql/core/assembler/TestInMemDatasetAssembler.java
@@ -50,7 +50,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 /**
- * Tests for {@link InMemDatasetAssembler}
+ * Tests for {@link TxnInMemDatasetAssembler}
  *
  */
 public class TestInMemDatasetAssembler extends Assert {
@@ -60,7 +60,7 @@ public class TestInMemDatasetAssembler extends Assert {
 	    model.setNsPrefix("ja", JA.getURI()) ;
 //	    System.out.println("-------------");
 //	    RDFDataMgr.write(System.out, model, Lang.TTL) ;
-	    final InMemDatasetAssembler testAssembler = new InMemDatasetAssembler();
+	    final TxnInMemDatasetAssembler testAssembler = new TxnInMemDatasetAssembler();
 		return testAssembler.open(testAssembler, example, DEFAULT);
 	}
